+++ 01libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: arblanco <arblanco@student.42urduli>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/20 17:40:36 by arblanco          #+#    #+#             */
-/*   Updated: 2023/01/18 16:58:42 by arblanco         ###   ########.fr       */
+/*   Updated: 2023/02/11 19:06:22 by arblanco         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
